requirement.txt
langchain
langchain-groq
faiss-cpu
tiktoken
langchain-community
langchain-openai
chromadb
sentence-transformers
pypdf
python-dotenv
pandas
matplotlib
langchain-huggingface
langchain-chroma
rank_bm25
docling
streamlit
tabulate